feat(exception): implement ExceptionAdvice for exception handling
- implement custom exception type for 404 NOT FOUND
