+++ src/main/java/wooteco/subway/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
