+++ src/main/java/wooteco/subway/ui/ExceptionAdvice.java
@@ -1,0 +1,27 @@
+package wooteco.subway.ui;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import wooteco.subway.exception.NotFoundException;
+
+@ControllerAdvice
+public class ExceptionAdvice {
+
+    @ExceptionHandler
+    public ResponseEntity<String> handleException(Exception e) {
+        e.printStackTrace();
+        return new ResponseEntity<>("예상치 못한 문제가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<String> handleException(IllegalArgumentException e) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<String> handleException(NotFoundException e) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+    }
+}
