+++ src/main/java/wooteco/subway/ui/StationController.java
@@ -1,17 +1,20 @@
 package wooteco.subway.ui;
 
-import org.springframework.http.HttpStatus;
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
-
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 public class StationController {
@@ -38,16 +41,5 @@ public class StationController {
     public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
         StationDao.deleteById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<String> handleException(Exception e) {
-        e.printStackTrace();
-        return new ResponseEntity<>("예상치 못한 문제가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<String> handleException(IllegalArgumentException e) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
     }
 }
